refactoring structures/..
